federated_query/requirements.txt
openai
psycopg2-binary
transformers
sentencepiece
